Flesh out lifecycle management of tunnel client
Since tunnel setup is more synchronous, we are actually establishing the
streams before the WebSocket handshake completes (because if we wait,
we'll receive the initial packets before we have destinations for them)

So things like conditional behavior off the handshaked protocol version
must be done lazily, while port forward setup must be done immediately
for websockets (and only websockets).

Not really looking forward to reconciling this code for whenever
Kubernetes introduces a better way to do port forwards over WebSocket.